Start testing the instruction parsing
This is a test that just extracts all of the instructions for an ISA and
deepseqs the result to make sure there are no errors lurking in the parsed
values.

This change includes ISA definitions for all of the supported ISAs (all
incomplete).  These ISA definitions have ISA-specific interpretations of various
bits of data.