Push AVR table construction through
This required a handy tweak to let us elide certain instructions from the
decoder tables.  This is primarily an issue around instructions that are
actually aliases handled by the assembler, but have identical bit sequences.
Ideally, we'll be able to pretty print them with the nicer mnemonics during
pretty printing, but that is future work.

This uncovered a bug in the LLVM AVR defs (separate commit coming)