Fix operand ordering during disassembly.
The disassembler now works for a bunch of instructions!

Remaining items:

* Fix endianness of assembler
* Fix operand assembly
* Fix pretty printing (especially of memory refs)

This change is based on what seems to be a convention in tablegen files relating
to how field descriptions are ordered, and how to map those to the DAG items.
There is some documentation in this commit.  This lets us remove that horrible
hack attempting to specify the mapping manually, which was essentially
impossible because it was inconsistent.  Operand orders now appear to be correct
from some spot checking.