Revise operand mapping
The previous attempt at the mapping relied on an order property that didn't seem
to be true.  It also could not accommodate operands that were defined by two
fields in the Inst definition.

The new approach is based on a plain argument name mapping, with the ability to
provide a custom mapping for each instruction form.