dismantle-tablegen: distinguish between input and tgen endianness
Previously we had isaEndianness. This was ostensibly supposed to
describe the ISA endianness, but was used to decide whether to
convert the endianness of the incoming Tgen bit patterns. This
ended up leading to undesirable behavior for ARM: a specification
of Big for the ARM endianness assumed that the Tgen patterns were
little-endian, which was not true, and then the endianness of those
patterns got flipped when they were processed.

This change splits up the endianness specification into two halves:
the input data endianness (for instruction parsing) and the Tgen
bit pattern endianness. We only convert the bit patterns when their
endianness is different from that of the incoming data to be parsed.