dismantle-tablegen: replace tgen endianness checking with bit pattern preprocessing
This change makes it so that instead of making ISAs declare the
endianness of the Tgen bit patterns, we require ISAs to specify a bit
pattern transformation to be applied at parse time. This is because in
practice, tgen bit patterns vary not by endianness (they are all
effectively Big) but instead vary by whether the leftmost bit in the
pattern indicates the highest bit or the lowest. The PPC tgen data
provides its instruction bit patterns with the least significant
(lowest) bit first in the list, while ARM specifies patterns with the
most significant bit first in the list.

Ultimately the Haskell code needs the most significant bit to be
specified first as a matter of normalization, so the transformation
function specified by the ISA is likely to be 'id' or
'Data.List.reverse' depending on the incoming tgen patterns.

Prior to this change we had some bit index reversal bugs which, when
combined with the honest endianness checking we were doing, made PPC
work by accident. Now, in addition to no longer making a tgen endianness
distinction, the table generation process assumes (as mentioned above)
that incoming bit patterns (post-transformation) provide their MSB first
and LSB last.