Rewrite the `ByteTrie` to support generating parsing tables at compile time
The previous `ByteTrie` implementation required the parsing tables to be
constructed at run-time, as preserving the sharing efficiently was very painful
to do with TH.

The new implementation flattens out the representation of the `ByteTrie` into
two arrays that are much easier to represent at compile time.  This should be
about as efficient and significantly improve startup times.