dismantle-tablegen: support ignoring per-file addresses during testing
This change makes it so that we can tell the testing procedure to skip
specific addresses in specific files. We do this by making the test
config specify a list of addresses to be ignored for a given file.
During test construction, any address in a file that is in the ignore
list is skipped.

This patch also provides initial ignore lists for PPC and ARM.