Add a TH helper to capture instance dictionaries
We need this to generate lists of `Some (Witness k Opcode)` for some class `k`.
Writing the list out by hand is impractical due to the sheer number of
constructors.  It doesn't seem possible to really defer the capture of these
dictionaries via trickery, so we'll just use TH for the job.

It is up to the caller to fix the type of `k` by adding a type signature.

This change includes moving the `Witness` type into this package (from semmc)
since we need to reference it here.