Fix a bug in `parseInstruction`
If the correct parser could be determined by inspecting only a prefix of the
bytes of the instruction, `parseInstruction` would return a short byte count.
For example, if it only needed to inspect two bytes to determine it needed to
decode the `ADDI` instruction, it would parse that instruction and return 2
bytes instead of 4.

While that was incorrect, it could also lead to a crash because the parser never
checked if there were really enough bytes (and then called partial functions).

The fix is to carry the expected number of actual bytes in the `Parser` data
type, and to check for the necessary bytes before invoking the parser.