Fix buggy negation for `W` bit vectors.
By changing to using `Integer` internally, so that `complement` is
defined, and then using explicit 2's complement. Because we don't want
negative ints inside, we hide the constructor and only export a smart
constructor `w`. With ViewPatterns this is not much different than
using the actual constructor `W`.

Alternatively, we could have just kept using `Natural`, and just
converted to `Integer` internally inside the implementation of the
operations that need `complement`. Thought of that just now while
committing, or I would have just done it that way in the first place.