Add a TH helper to generate some `OrdF` and `TestEquality` instances
These unfortunately can't be generated in the main `genISA` call because the
helpers from parameterized-utils call a variant of `reify`, which requires the
referenced type to be in the AST already.  This means we have to split the TH
calls with a `$(return [])` to force TH to flush its state and make those
definitions available.  The PPC architecture implementation shows how this
should be handled.