Add a new dictionary capture helper
The new helper also captures some extra information for each opcode so that we
can generate some TH for them later.  The extra information includes the TH Name
of each opcode, as well as a bit of information to generate operand list
matchers for each opcode.

This commit also moves the capture helpers to their own modules, since they are
really not related to the other TH code.