Support Tgen data that uses halfword bit patterns
This change modifies tablegen to support Tgen data whose bit patterns
are full word length (32 bits) even though only some of the entries in
the pattern are actually used (e.g. 16 bits in the case of thumb). To do
this, we rely on the presence of a "Size" field (in bytes) of the Tgen
instruction descriptor to indicate how many pattern entries are actually
used. But that isn't enough: we now know how many entries to use, but
not which entries (the first 16? the last 16?). To answer that question
we now allow ISA implementors to provide an ISA configuration setting,
isaUnusedBitsPolicy. This setting is optional but takes the values
"Take" or "Drop" if provided. "Take" means that any used bits in the
pattern are taken from the beginning; "Drop" means that they are taken
from the end.