tablegen: support filtering instructions prior to test case generation
This change makes it so that ArchTestConfigs provide an
instructionFilter function to filter instructions prior to generating
test cases. This allows some architecture tests to ignore some
instructions altogether, which is not the same as expecting them to fail
(which would have been another way to implement this feature).

This change also makes the objdump output parser use more specific
parsers for the different formats of instruction byte sequences. The
parser now distinguishes between different byte layouts and preserves
that information in the parsed Instructions so that they can be filtered
on that basis. Concretely, this is used by the Thumb tests to ensure
that any full-word ARM instructions are ignored in the tests since many
Thumb binaries also contain ARM instructions that will fail parsing in
the Thumb disassembler.