tablegen: fix a format string parsing bug
This fixes a bug in which a format string uses two adjacent unbracketed
variables, e.g. "$foo$bar". Before the fix, "$foo" would correctly get
interpolated but the "$" from "$bar" would be preserved in the final
string and "bar" would go on to be interpreted as literal text. That was
broken because the "$" should have been restored into the format string
for the next formatting pass. This fixes that.