Generate constructors for all operand types in the ISA
Instead of just the operand types observed in the .tgen file. We still
also attempt to generate all observed operand types, so we'll still
get a helpful error if the .tgen file contains and operand type not
specified in the ISA. This change has no effect on production where we
use the full .tgen file, where all operand types defined in the ISA
will occur in the .tgen file. But this change matters for building a
against minimal .tgen file, where not all operand types will occur.