dismantle-tablegen: add binary representation to pretty print comparison failures
When pretty print comparisons fail, it's very difficult to determine
which operand types and tablegen descriptor are associated with the
failure to reproduce the objdump output. This change includes the binary
representation of the input bytes in the pretty print comparison test
failure messages so that the tablegen can be matched to more easily
determine both the descriptor and operand types to blame for the
failure.