Experimental: TH: split up assembler case expression
This change is a compilation performance experiment.

Previously, the TH-generated assembler used a case expression to match
on instructions to determine how to assemble them. This experiment was
to determine how compilation time would be affected if we were to split
up the case expression so that instead of matching on case branches, we
perform a table lookup on the instruction opcode and dispatch to a
separate top-level function to perform the assembly.

The result of this experiment was as follows. The times indicate the
time to build the dismantle-arm package:

Before the change:

real    2m52.312s
user    2m50.640s
sys     0m1.880s

After the change:

real    2m30.284s
user    2m28.800s
sys     0m1.776s