Experimental: TH: split up pretty printer case expression
This change is a compilation performance experiment.

Previously, the TH-generated pretty printer used a case expression to
match on instructions to determine how to print them. This experiment
was to determine how compilation time would be affected if we were to
split up the case expression similar to the previous commit. This change
also led to a significant reduction in build times for both the ARM and
AArch64 packages.