dismantle: support custom objdump arguments on a per-binary basis
This change makes it so that testing configurations can specify custom
arguments to objdump when disassembling test binaries. Although this
isn't used in this commit, it paves the way for disassembling non-elf
binaries during testing. In those cases objdump needs architecture hints
about the binary in question and requires a different disassembly flag.
Any binary without a custom objdump argument list in the testing
configuration will just use the default as before.