Remove some KnownNat constraints
We now include NatReprs in the W length-indexed word type so that we don't
require KnownNat constraints in as many places.  This is required so that we can
get rid of uses of withKnownNat in some downstream code, as that function is
going away.

Many helpers retain the same names/instances with the constraint, but new
functions have been added to enable working with W values without the KnownNat
constraint.  The old instances and functions are left in place to allow simpler
use in contexts where we really do have a KnownNat constraint (e.g., in the TH
generated code).