[arm] use safe fromIntegral for creating W.W values.
The fromIntegral will generate a runtime error if the input value
exceeds the size supported by the 'W.W n' value, whereas the 'W.w'
constructor (the use of which was removed in this commit) will discard
excess bits silently.  The 'W.w' behavior is dangerous
   (e.g. '(47 == (W.w 1 :: W.W 1)) == True')
and should only be used when explicitly known to be appropriate.