[tablegen] functional testing for ByteTrie building and lookup.
Note that this testing uncovered some possibly questionable behavior
of the ByteTrie.  The tests are currently set to accept this behavior
in order to pass, but these should be examined to determine if the
current behavior is actually desireable or if it should be changed.
These behaviors are marked in the tests with the word "QUESTIONABLE".