Initial draft of a new BitTrie structure
This is very similar to the ByteTrie, but doing dispatch at the bit level
instead.  The externally-facing API is identical.  The goal of this new
structure is to have better space use in "dense" encodings (like Thumb) where
the radix tree approach is too inefficient.

This commit includes a bit of refactoring to ensure that we can share the
underlying data structure and unsafe TH-oriented primitives between the two
implementations.