Fix some more logic around negative pattern matching
This commit is more faithful to the semantics of negative pattern matching.
Negative patterns can only reject a matching bitstring if the bitstring is long
enough to include *all* of the negative bits in the pattern.

It also fixes some of the matching logic to not only give early matching
preference to short patterns that fully cover the bit string, but also filters
out shorter paths once the bit string grows larger than the pattern.

This commit rewrites the `makePayload` function to attempt to make its logic
easier to follow.