Update Trie tests to check for full pattern match.
Previous behavior was to match only up to the first unique byte and
ignore the rest of the pattern.

Note: this adds a depth validation check that results in a very large
number of tests being run.  To debug other test failures, it may help
to locate the two testgroups labelled with "MANYTESTS!!" and comment
them out.