Progress on caching in the BitTrie
NOTE: This does not completely work yet.  It caches enough to make building the
T32 decoder possible (~10GB RAM during compilation).  There are a few failing
tests, but not many, suggesting that this change might be salvageable.

The basic idea is that we cache everything *except* the "spine" of the decoder
that corresponds to an exact match to a negative pattern (caching across
negative patterns can yield incorrect parse tables).