Introduce a tool to compute parse tables offline
This serializes the computed Trie for a given set of inputs.  The idea is that
we can check these in (since they are expensive to compute for some
architectures) and use the precomputed tables from Template Haskell.

To ensure consistency, we embed a hash of all of the inputs in the serialized
parse tables.