Implement a new parse table structure
This commit replaces the ByteTrie, which formed the core of the disassembler
infrastructure, with a BitTrie.  Instead of parsing byte-by-byte, we now parse
bit-by-bit.  The underlying motivation was to reduce compile times and memory
requirements for the ASL-based ARM disassembler (especially for Thumb).

In diagnosing the memory requirements of the Thumb disassembler, it became
obvious that the relative density of the instruction set combined with the
specification by means of negative patterns simply produced enormous parsers
without much sharing at the leaves.  The generated parse tables where about 3-4x
the size of the A32 disassembler.  The density and lack of sharing was
exacerbated by the operand boundaries of T32 instruction encodings frequently
spanning byte boundaries, which destroyed sharing.

By switching to a bit-level tire structure, we can get nearly optimal
sharing (at the expense of a bit of decode speed).  As a side effect of the
changes, the generated parse tables are significantly smaller.

Unfortunately, during the implementation of the BitTrie, it became apparent that
the caching strategy used for the ByteTrie was unsafe in the presence of
negative patterns and was only mostly working by coincidence.  The caching code
has been significantly rewritten to support the BitTrie.

Fixing the caching ended up making the time required to generate the parse
tables unreasonable (though memory use is now negligible).  To mitigate this,
there is now a tool (dismantle-arm-xml-tablegen) that pre-computes the parse
tables and persists them on disk.  They are checked into the repository.

The commit is wide-reaching and unfortunately fixes a few unrelated things at
the same time.  Breaking them into separate commits would have caused quite a
bit of extra work.  Summary of changes:
- Always examine all instruction bits before committing to a parse (in case
  there are negative patterns that would match the last bit)
- Pre-compute parse tables for dismantle-arm-xml to reduce compile times
- Fixes caching to respect negative patterns
- Fixes the original dismantle-arm-xml tests for the "lite" set. They were
  missing an instruction, which was masked by an incorrect decoding (ADR
  mistakenly parsed as ADDI)
- Simplifies the table generation protocol (specifically payload storage) to
  avoid the need for an `undefined` padding element
- Replace the ByteTrie with the new BitTrie; ByteTrie has been removed
- Simplifies and documents the pattern (and negative pattern) matching logic

Future work:
- Identify a more efficient table generation strategy

Fixes #18, Fixes #16, Fixes #14