Enable a Windows build (8.6.5) (#25)
Re-enable the Windows build in CI

The CI caught a problem: line ending translation on Windows perturbs the inputs to the decoder generator, which means that the hashes in the pre-generated parse tables fail to match when we check consistency.  The fix is to filter out all newline characters when computing the hashes.