WIP Fixes to the Thumb decoder
The original implementation generated bit masks that were too permissive (and
too large) that caused the decoder to improperly consume 16 extra bits with
every T16 encoded instruction.

This (possibly misguided) attempt to fix it type-indexes all of the bitmasks to
ensure that we get it right.

Note: it currently fails at TH time due to an error somewhere.