Fixes to the Thumb decoder (#30)
Fix the Thumb decoder

The original implementation always stored 4 byte required patterns and masks, even for two byte instructions. These happened to create an unconstrained two-byte prefix that meant that decoding a Thumb instruction always consumed the instruction before it, which incorrectly dropped half of the thumb instructions.

This change explicitly tracks the width of every instruction encoding (using a `NatRepr`). While the change is large by lines-of-code, it is really just pushing `NatRepr`s around to make the widths of instruction encodings explicit everywhere. This change uncovered the places making the incorrect assumptions about instructions being 4 bytes everywhere (there was an attempt at adding the correct mask in the T16 case, but the resulting mask was the incorrect one).

There are a few places where instruction field (operand) bit indices had to be corrected, since they are all counted from 31 even on 16 bit instructions. These have been marked with comments to try to make them obvious, as they are a bit tricky.

Important note: 4 byte Thumb instructions (i.e., Thumb2) are not quite working. It seems to be an endianness issue (which might be mostly a problem with the test suite). However, as it works well with Thumb1, it is a major improvement nonetheless.

Additional changes included that were required to make Thumb work:
- This change also includes some more detailed error reporting (including contexts for errors that can be raised in more than one place).
- This change disables the encoding validator. That code attempted to confirm that the encodings extracted from the ASL spec matched the encodings suggested by the separate encodings table (which is a separate XML file). This check meant well, but the names of many Thumb instruction encodings don't line up cleanly between the two data sources, so most Thumb instructions were failing only because we didn't have a good mapping. We now just assume that the ASL is canonical and correct. If we find that there are substantial errors due to discrepancies, we can restore this check, but it will require some effort to divine the mapping between the two data sources.
- This change uses `nub` to make negative patterns unique in each encoding. There were some errors in the Thumb translation log due to some negative patterns not quite matching up