Fix the parsing of Thumb2 instructions (#35)
See the details in the `endianness` function documentation. Long story short:
Thumb2 instructions are parsed as two independent 16 bit words (rather than a
single Word32), which requires different endian swapping.

Note that all of the Thumb2 parse tables were incorrect before. This change
significantly shrank the pre-generated Thumb parsing tables.

The test harness needed a fix to test this correctly, as it had the same Thumb2
endian swapping bug as the actual code (but in a slightly different
manifestation).